# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Life Circle Backend",
    "version": "2.0.0"
  },
  "paths": {
    "/api/analyses/{task_id}/routes/{facility_id}": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Facility Route",
        "operationId": "facility_route_api_analyses__task_id__routes__facility_id__post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Facility Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteEvidence"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/checkups": {
      "post": {
        "tags": [
          "checkups"
        ],
        "summary": "Create",
        "operationId": "create_api_v2_checkups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckupTaskView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/checkups/sessions": {
      "post": {
        "tags": [
          "checkups"
        ],
        "summary": "Open Session",
        "operationId": "open_session_api_v2_checkups_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionOpenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/checkups/sessions/{session_id}/tabs/{tab_id}/heartbeat": {
      "post": {
        "tags": [
          "checkups"
        ],
        "summary": "Heartbeat",
        "operationId": "heartbeat_api_v2_checkups_sessions__session_id__tabs__tab_id__heartbeat_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "tab_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tab Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/checkups/sessions/{session_id}/tabs/{tab_id}/close": {
      "post": {
        "tags": [
          "checkups"
        ],
        "summary": "Close Tab",
        "operationId": "close_tab_api_v2_checkups_sessions__session_id__tabs__tab_id__close_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "tab_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tab Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/checkups/by-request/{client_request_id}": {
      "get": {
        "tags": [
          "checkups"
        ],
        "summary": "By Request",
        "operationId": "by_request_api_v2_checkups_by_request__client_request_id__get",
        "parameters": [
          {
            "name": "client_request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Client Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckupTaskView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/checkups/{task_id}": {
      "get": {
        "tags": [
          "checkups"
        ],
        "summary": "Status",
        "operationId": "status_api_v2_checkups__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckupTaskView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/checkups/{task_id}/result": {
      "get": {
        "tags": [
          "checkups"
        ],
        "summary": "Result",
        "operationId": "result_api_v2_checkups__task_id__result_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckupSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/checkups/{task_id}/layers/{layer_id}": {
      "get": {
        "tags": [
          "checkups"
        ],
        "summary": "Layer",
        "operationId": "layer_api_v2_checkups__task_id__layers__layer_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RouteEvidence": {
        "properties": {
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "endpoint_verified": {
            "type": "boolean",
            "title": "Endpoint Verified"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "path": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Path"
          },
          "poiEvidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PoiEvidence"
              },
              {
                "type": "null"
              }
            ],
            "x-legacy-optional": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "distance_m",
          "duration_s",
          "endpoint_verified",
          "reason",
          "path"
        ],
        "title": "RouteEvidence"
      },
...
    }
  }
}